@@ -16,12 +16,13 @@
     "@angular/common": "~8.1.1",
     "@angular/compiler": "~8.1.1",
     "@angular/core": "^8.0.3",
+    "@angular/flex-layout": "~8.0.0-beta.27",
     "@angular/forms": "~8.1.1",
     "@angular/material": "^8.0.1",
     "@angular/platform-browser": "~8.1.1",
     "@angular/platform-browser-dynamic": "~8.1.1",
     "@angular/router": "~8.1.1",
-    "core-js": "^2.5.4",
+    "core-js": "^2.6.12",
     "hammerjs": "^2.0.8",
     "ngx-material-timepicker": "^4.0.2",
     "rxjs": "~6.5.2",
@@ -35,7 +36,10 @@
     "@angular/language-service": "~8.1.1",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
+    "@types/node": "^10.17.48",
+    "angular2-template-loader": "^0.6.2",
     "codelyzer": "^5.0.1",
+    "html-webpack-plugin": "^3.2.0",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~4.0.0",
@@ -44,15 +48,12 @@
     "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
     "protractor": "~5.4.0",
+    "raw-loader": "^0.5.1",
+    "ts-loader": "^5.2.2",
     "ts-node": "~7.0.0",
     "tslint": "~5.11.0",
     "typescript": "~3.4.5",
-    "@types/node": "^10.12.2",
-    "angular2-template-loader": "^0.6.2",
-    "html-webpack-plugin": "^3.2.0",
-    "raw-loader": "^0.5.1",
-    "ts-loader": "^5.2.2",
-    "webpack": "^4.24.0",
+    "webpack": "^4.44.2",
     "webpack-cli": "^3.1.2",
     "webpack-dev-server": "^3.1.10"
   }
